Gelman-Rubin criterion for convergence

Gelman-Rubin criterion for convergence (uses gelman.diag)

GR_crit(object, confidence = 0.95, transform = FALSE, autoburnin = TRUE,
  multivariate = TRUE, subset = NULL, start = NULL, end = NULL,
  thin = NULL, warn = TRUE, ...)

Arguments

object
object inheriting from class JointAI

confidence
the coverage probability of the confidence interval for the potential scale reduction factor

transform
a logical flag indicating whether variables in x should be transformed to improve the normality of the distribution. If set to TRUE, a log transform or logit transform, as appropriate, will be applied.

autoburnin
a logical flag indicating whether only the second half of the series should be used in the computation. If set to TRUE (default) and start(x) is less than end(x)/2 then start of series will be adjusted so that only second half of series is used.

multivariate
a logical flag indicating whether the multivariate potential scale reduction factor should be calculated for multivariate chains

subset
subset of parameters/variables/nodes (columns in the MCMC sample). Uses the same logic as the argument monitor_params in lm_imp, glm_imp and lme_imp.

start
the first iteration of interest (see window.mcmc)

end
the last iteration of interest (see window.mcmc)

thin
thinning interval (see window.mcmc)

warn
logical; should warnings be given?; default is TRUE. Note: this applies only to warnings given directly by JointAI.

...
currently not used

References

Gelman, A., Meng, X. L., & Stern, H. (1996). Posterior predictive assessment of model fitness via realized discrepancies. Statistica Sinica, 733-760.

See also

Examples

mod1 <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)

#> This is new software. Please report any bug to the package maintainer.

#>   |                                                          |                                                  |   0%  |                                                          |*                                                 |   2%  |                                                          |**                                                |   4%  |                                                          |***                                               |   6%  |                                                          |****                                              |   8%  |                                                          |*****                                             |  10%  |                                                          |******                                            |  12%  |                                                          |*******                                           |  14%  |                                                          |********                                          |  16%  |                                                          |*********                                         |  18%  |                                                          |**********                                        |  20%  |                                                          |***********                                       |  22%  |                                                          |************                                      |  24%  |                                                          |*************                                     |  26%  |                                                          |**************                                    |  28%  |                                                          |***************                                   |  30%  |                                                          |****************                                  |  32%  |                                                          |*****************                                 |  34%  |                                                          |******************                                |  36%  |                                                          |*******************                               |  38%  |                                                          |********************                              |  40%  |                                                          |*********************                             |  42%  |                                                          |**********************                            |  44%  |                                                          |***********************                           |  46%  |                                                          |************************                          |  48%  |                                                          |*************************                         |  50%  |                                                          |**************************                        |  52%  |                                                          |***************************                       |  54%  |                                                          |****************************                      |  56%  |                                                          |*****************************                     |  58%  |                                                          |******************************                    |  60%  |                                                          |*******************************                   |  62%  |                                                          |********************************                  |  64%  |                                                          |*********************************                 |  66%  |                                                          |**********************************                |  68%  |                                                          |***********************************               |  70%  |                                                          |************************************              |  72%  |                                                          |*************************************             |  74%  |                                                          |**************************************            |  76%  |                                                          |***************************************           |  78%  |                                                          |****************************************          |  80%  |                                                          |*****************************************         |  82%  |                                                          |******************************************        |  84%  |                                                          |*******************************************       |  86%  |                                                          |********************************************      |  88%  |                                                          |*********************************************     |  90%  |                                                          |**********************************************    |  92%  |                                                          |***********************************************   |  94%  |                                                          |************************************************  |  96%  |                                                          |************************************************* |  98%  |                                                          |**************************************************| 100%

GR_crit(mod1)

#> Potential scale reduction factors:
#> 
#>             Point est. Upper C.I.
#> (Intercept)      0.999       1.01
#> C1               0.999       1.01
#> M22              1.078       1.26
#> M23              1.061       1.21
#> M24              1.071       1.24
#> C2               1.000       1.01
#> tau_y            1.009       1.04
#> sigma_y          1.011       1.05
#> 
#> Multivariate psrf
#> 
#> 1.08