JointAI: Joint Analysis and Imputation of Incomplete Data

The JointAI package performs simultaneous imputation and inference for incomplete data using the Bayesian framework. Distributions for incomplete variables are specified automatically and modeled jointly with the analysis model.

Main functions

The package has three main functions, lm_imp, glm_imp and lme_imp, that allow analysis using linear regression, generalized linear regression and linear mixed effects models. As far as possible, the specification of the functions is the same as the specification of their complete data versions lm, glm and lme (from the package nlme).

Results can be summarized and printed with summary.JointAI, and visualized using traceplot or densplot. The function predict.JointAI allows prediction (including credible intervals) from JointAI models.

Evaluation and export

Two criteria for evaluation of convergence and precision of the posterior estimate are available: GR_crit and MC_error.

Imputed data can be extracted and exported to SPSS data using get_MIdat.

Vignettes

The following vignettes are available

Minimal Example:
A minimal example demonstrating the use of lm_imp, summary.JointAI, traceplot and densplot.

Visualizing Incomplete Data:
Demonstrations of the options in plot_all (plotting histograms and barplots for all variables in the data) and md_pattern (plotting or printing the missing data pattern).

Model Specification:
Explanation and demonstration of all parameters that are required or optional to specify the model structure in lm_imp, glm_imp and lme_imp. Among others, the functions parameters, list_impmodels, get_imp_meth and set_refcat are used.

Parameter Selection:
Examples on how to select the parameters/variables/nodes to follow using the argument monitor_params and the parameters/variables/nodes displayed in the summary, traceplot, densplot or when using GR_crit or MC_error.

References

Erler, N. S., Rizopoulos, D., Rosmalen, J. V., Jaddoe, V. W., Franco, O. H., & Lesaffre, E. M. (2016). Dealing with missing covariates in epidemiologic studies: A comparison between multiple imputation and a full Bayesian approach. Statistics in Medicine, 35(17), 2955-2974.