National Health and Nutrition Examination Survey (NHANES) Data

This data is a small subset of the data collected within the 2011--2012 wave of the NHANES study, a study designed to assess the health and nutritional status of adults and children in the United States, conduced by the National Center for Health Statistics.

data(NHANES)

Format

A data frame with 198 rows and 11 variables:

SBP

systolic blood pressure

gender

male or female

age

in years

race

race / hispanic origin (5 categories)

WC

waist circumference in cm

alc

alcohol consumption (binary: <1 drink per week vs. >= 1 drink per week)

educ

educational level (binary: low vs. high)

creat

creatinine concentrationin in mg/dL

albu

albumin concentration in g/dL

uricacid

uric acid concentration in mg/dL

bili

bilirubin concentration in mg/dL

Source

National Center for Health Statistics (NCHS) (2011 - 2012). National Health and Nutrition Examination Survey Data. URL https://www.cdc.gov/nchs/nhanes/.

Note

The subset provided here was selected and re-coded to facilitate demonstration of the functionality of the JointAI package, and no clinical conclusions should be derived from it.

Examples

summary(NHANES)

#>       SBP            gender         age                        race   
#>  Min.   : 93.33   male  :119   Min.   :20.00   Mexican American  :26  
#>  1st Qu.:117.33   female: 79   1st Qu.:35.00   Other Hispanic    :24  
#>  Median :127.33                Median :43.50   Non-Hispanic White:71  
#>  Mean   :130.80                Mean   :44.81   Non-Hispanic Black:47  
#>  3rd Qu.:140.67                3rd Qu.:55.00   other             :30  
#>  Max.   :222.00                Max.   :78.00                          
#>                                                                       
#>        WC          alc       educ         creat             albu      
#>  Min.   : 70.5   <1  :76   low :112   Min.   :0.4000   Min.   :3.200  
#>  1st Qu.: 93.2   >=1 :75   high: 86   1st Qu.:0.7150   1st Qu.:4.100  
#>  Median :102.4   NA's:47              Median :0.8400   Median :4.300  
#>  Mean   :102.8                        Mean   :0.8472   Mean   :4.299  
#>  3rd Qu.:111.3                        3rd Qu.:0.9400   3rd Qu.:4.500  
#>  Max.   :151.0                        Max.   :2.3000   Max.   :5.200  
#>  NA's   :8                            NA's   :15       NA's   :15     
#>     uricacid          bili       
#>  Min.   :2.500   Min.   :0.2000  
#>  1st Qu.:4.600   1st Qu.:0.5000  
#>  Median :5.500   Median :0.6000  
#>  Mean   :5.534   Mean   :0.6956  
#>  3rd Qu.:6.400   3rd Qu.:0.8000  
#>  Max.   :9.100   Max.   :1.8000  
#>  NA's   :15      NA's   :15