Get default values for hyperparameters

Prints the list of default values for the hyperparameters.

default_hyperpars(family = "gaussian", link = "identity",
  nranef = NULL)

Arguments

family
distribution family of the analysis model (gaussian, binomial, poisson or Gamma)

link
link function (if the link is already given in the family, e.g. family = binomial("logit")) this argument does not need to be specified

nranef
number of random effects

Value

A list containing the default hyperparameters for JointAI models. The elements of the list are

analysis_model: hyperparameters for the analysis model

mu_reg_main

mean in the priors for regression coefficients

tau_reg_main

precision in the priors for regression coefficients

a_tau_main

scale parameter in Gamma prior for precision of outcome

b_tau_main

rate parameter in Gamma prior for precision of outcome

Z: hyperparameters for the random effects in mixed models

RinvD

scale matrix in Wishart prior (*) for random effects covariance matrix

KinvD

degrees of freedom in Wishart prior for random effects covariance matrix

a_diag_RinvD

scale parameter in Gamma prior for the diagonal elements of

RinvD

b_diag_RinvD

rate parameter in Gamma prior for the diagonal elements of

RinvD

(*) when there is only one random effect a Gamma distribution is used instead of the Wishart

norm: hyperparameters for normal and lognormal imputation models

mu_reg_norm

mean in the priors for regression coefficients

tau_reg_norm

precision in the priors for regression coefficients

a_tau_norm

scale parameter in Gamma prior for precision of imputed variable

b_tau_norm

rate parameter in Gamma prior for precision of imputed variable

gamma: hyperparameters for Gamma imputation models

mu_reg_gamma

mean in the priors for regression coefficients

tau_reg_gamma

precision in the priors for regression coefficients

a_tau_gamma

scale parameter in Gamma prior for precision of imputed variable

b_tau_gamma

rate parameter in Gamma prior for precision of imputed variable

beta: hyperparameters for beta imputation models

mu_reg_beta

mean in the priors for regression coefficients

tau_reg_beta

precision in the priors for regression coefficients

a_tau_beta

scale parameter in Gamma prior for precision of imputed variable

b_tau_beta

rate parameter in Gamma prior for precision of imputed variable

logit: hyperparameters for logistic imputation models

mu_reg_logit

mean in the priors for regression coefficients

tau_reg_logit

precision in the priors for regression coefficients

multinomial: hyperparameters for multinomial imputation models

mu_reg_multinomial

mean in the priors for regression coefficients

tau_reg_multinomial

precision in the priors for regression coefficients

ordinal: hyperparameters for ordinal imputation models

mu_reg_ordinal

mean in the priors for regression coefficients

tau_reg_ordinal

precision in the priors for regression coefficients

mu_delta_ordinal

mean in the prior for the intercepts

tau_delta_ordinal

precision in the priors for the intercepts

Examples

default_hyperpars()

#> $analysis_model
#>  mu_reg_main tau_reg_main   a_tau_main   b_tau_main 
#>        0e+00        1e-04        1e-02        1e-03 
#> 
#> $Z
#> NULL
#> 
#> $norm
#>  mu_reg_norm tau_reg_norm   a_tau_norm   b_tau_norm 
#>        0e+00        1e-04        1e-02        1e-02 
#> 
#> $gamma
#>  mu_reg_gamma tau_reg_gamma   a_tau_gamma   b_tau_gamma 
#>         0e+00         1e-04         1e-02         1e-02 
#> 
#> $beta
#>  mu_reg_beta tau_reg_beta   a_tau_beta   b_tau_beta 
#>        0e+00        1e-04        1e-02        1e-02 
#> 
#> $logit
#>  mu_reg_logit tau_reg_logit 
#>         0.000         0.001 
#> 
#> $multinomial
#>  mu_reg_multinomial tau_reg_multinomial 
#>               0.000               0.001 
#> 
#> $ordinal
#>    mu_reg_ordinal   tau_reg_ordinal  mu_delta_ordinal tau_delta_ordinal 
#>             0.000             0.001             0.000             0.001 
#>