Parameters used by several functions in JointAI

Parameters used by several functions in JointAI

Arguments

object

object inheriting from class 'JointAI'

no_model

optional; vector of names of variables for which no model should be specified. Note that this is only possible for completely observed variables and implies the assumptions of independence between the excluded variable and the incomplete variables.

timevar

name of the variable indicating the time of the measurement of a time-varying covariate in a proportional hazards survival model (also in a joint model). The variable specified in "timevar" will automatically be added to "no_model".

assoc_type

named vector specifying the type of the association used for a time-varying covariate in the linear predictor of the survival model when using a "JM" model. Implemented options are "underl.value" (linear predictor; default for covariates modelled using a Gaussian, Gamma, beta or log-normal distribution) covariates) and "obs.value" (the observed/imputed value; default for covariates modelled using other distributions).

subset

subset of parameters/variables/nodes (columns in the MCMC sample). Follows the same principle as the argument monitor_params in *_imp.

exclude_chains

optional vector of the index numbers of chains that should be excluded

start

the first iteration of interest (see window.mcmc)

end

the last iteration of interest (see window.mcmc)

n.adapt

number of iterations for adaptation of the MCMC samplers (see adapt)

n.iter

number of iterations of the MCMC chain (after adaptation; see coda.samples)

n.chains

number of MCMC chains

quiet

logical; if TRUE then messages generated by rjags during compilation as well as the progress bar for the adaptive phase will be suppressed, (see jags.model)

progress.bar

character string specifying the type of progress bar. Possible values are "text" (default), "gui", and "none" (see update). Note: when sampling is performed in parallel it is not possible to display a progress bar.

thin

thinning interval (integer; see window.mcmc). For example, thin = 1 (default) will keep the MCMC samples from all iterations; thin = 5 would only keep every 5th iteration.

nrow

optional; number of rows in the plot layout; automatically chosen if unspecified

ncol

optional; number of columns in the plot layout; automatically chosen if unspecified

use_ggplot

logical; Should ggplot be used instead of the base graphics?

warn

logical; should warnings be given? Default is TRUE.

mess

logical; should messages be given? Default is TRUE.

xlab, ylab

labels for the x- and y-axis

idvars

name of the column that specifies the multi-level grouping structure

seed

optional; seed value (for reproducibility)

ppc

logical: should monitors for posterior predictive checks be set? (not yet used)

rd_vcov

optional character string or list (of lists or character strings) specifying the structure of the variance covariance matrix/matrices of the random effects for multivariate mixed models. Options are "full, "blockdiag" (default) and "indep". Different structures can be specified per grouping level (in multi-level models with more than two levels) by specifying a list with elements per grouping level. To specify different structures for different outcomes, a list (maybe nested in the list per grouping level) can be specified. This list should have the type of structure as names and contain vectors of variable names that belong to the respective structure.