Simulated Longitudinal Data in Long and Wide Format

This data was simulated to mimic data from a longitudinal cohort study following mothers and their child from birth until approximately 4 years of age. It contains 3903 observations of 500 mother-child pairs. Children's BMI and head circumference was measured repeatedly and their age in months was recorded at each measurement. Furthermore, the data contain several baseline variables with information on the mothers' demographics and socioeconomic status.

simLong

simWide

Format

simLong: A data frame in long format with 3908 rows and 13 variables

simWide: A data frame in wide format with 500 rows and 45 variables

Baseline covariates

(in simLong and simWide)

GESTBIR

gestational age at birth (in weeks)

ETHN

ethnicity (binary: European vs. other)

AGE_M

age of the mother at intake

HEIGHT_M

height of the mother (in cm)

PARITY

number of times the mother has given birth (binary: 0 vs. >=1)

SMOKE

smoking status of the mother during pregnancy (3 ordered categories: never smoked during pregnancy, smoked until pregnancy was known, continued smoking in pregnancy)

EDUC

educational level of the mother (3 ordered categories: low, mid, high)

MARITAL

marital status (3 categories)

ID

subject identifier

Long-format variables

(only in simLong)

time

measurement occasion/visit (by design, children should be measured at/around 1, 2, 3, 4, 7, 11, 15, 20, 26, 32, 40 and 50 months of age)

age

child age at measurement time in months

bmi

child BMI

hc

child head circumference in cm

Wide-format variables

(only in simWide)

age1, age2, age3, age4, age7, age11, age15, age20, age26, age32, age40, age50

child age at the repeated measurements in months

bmi1, bmi2, bmi3, bmi4, bmi7, bmi11, bmi15, bmi20, bmi26, bmi32, bmi40, bmi50

repeated measurements of child BMI

age1, age2, age3, age4, age7, age11, age15, age20, age26, age32, age40, age50

repeated measurements of child head circumference in cm

Examples

 summary(simLong)

#>     GESTBIR            ETHN          AGE_M          HEIGHT_M      PARITY    
#>  Min.   :36.29   European:2503   Min.   :17.06   Min.   :146.1   0   :2364  
#>  1st Qu.:39.37   other   :1308   1st Qu.:27.03   1st Qu.:163.0   >= 1:1451  
#>  Median :40.33   NA's    :  97   Median :30.42   Median :168.1   NA's:  93  
#>  Mean   :40.24                   Mean   :30.50   Mean   :168.0              
#>  3rd Qu.:41.13                   3rd Qu.:34.11   3rd Qu.:173.1              
#>  Max.   :43.62                   Max.   :46.73   Max.   :192.5              
#>                                                  NA's   :75                 
#>                               SMOKE        EDUC                 MARITAL    
#>  never smoked during pregnancy   :2620   high:1748   married        :1792  
#>  smoked until pregnancy was known: 278   mid :1564   living together:1456  
#>  continued smoking in pregnancy  : 539   low : 328   no partner     : 431  
#>  NA's                            : 471   NA's: 268   NA's           : 229  
#>                                                                            
#>                                                                            
#>                                                                            
#>       time            age               bmi              hc       
#>  Min.   : 1.00   Min.   : 0.3119   Min.   :10.66   Min.   :32.58  
#>  1st Qu.: 4.00   1st Qu.: 4.3882   1st Qu.:15.51   1st Qu.:42.39  
#>  Median :15.00   Median :13.1977   Median :16.61   Median :46.05  
#>  Mean   :17.26   Mean   :16.4707   Mean   :16.58   Mean   :45.56  
#>  3rd Qu.:26.00   3rd Qu.:25.5121   3rd Qu.:17.64   3rd Qu.:48.95  
#>  Max.   :50.00   Max.   :49.2852   Max.   :22.38   Max.   :56.37  
#>                                                                   
#>       ID           
#>  Length:3908       
#>  Class :character  
#>  Mode  :character  
#>                    
#>                    
#>                    
#>                    

 summary(simWide)

#>     GESTBIR            ETHN         AGE_M          HEIGHT_M      PARITY   
#>  Min.   :36.29   European:313   Min.   :17.06   Min.   :146.1   0   :307  
#>  1st Qu.:39.31   other   :173   1st Qu.:26.79   1st Qu.:162.7   >= 1:181  
#>  Median :40.30   NA's    : 14   Median :30.06   Median :168.1   NA's: 12  
#>  Mean   :40.19                  Mean   :30.28   Mean   :167.8             
#>  3rd Qu.:41.09                  3rd Qu.:33.83   3rd Qu.:172.8             
#>  Max.   :43.62                  Max.   :46.73   Max.   :192.5             
#>                                                 NA's   :10                
#>                               SMOKE       EDUC                MARITAL   
#>  never smoked during pregnancy   :333   high:220   married        :226  
#>  smoked until pregnancy was known: 36   mid :197   living together:184  
#>  continued smoking in pregnancy  : 70   low : 44   no partner     : 55  
#>  NA's                            : 61   NA's: 39   NA's           : 35  
#>                                                                         
#>                                                                         
#>                                                                         
#>       ID                 age1             bmi1         hc_CB0t2m    
#>  Length:500         Min.   :0.3119   Min.   :11.32   Min.   :32.35  
#>  Class :character   1st Qu.:0.9588   1st Qu.:14.11   1st Qu.:36.12  
#>  Mode  :character   Median :1.1197   Median :15.15   Median :37.14  
#>                     Mean   :1.1239   Mean   :15.12   Mean   :37.21  
#>                     3rd Qu.:1.3019   3rd Qu.:16.08   3rd Qu.:38.22  
#>                     Max.   :1.7615   Max.   :19.43   Max.   :41.53  
#>                                      NA's   :175                    
#>       age2            bmi2          hc_CB02m          age3      
#>  Min.   :2.002   Min.   :11.72   Min.   :35.93   Min.   :3.001  
#>  1st Qu.:2.216   1st Qu.:15.29   1st Qu.:39.24   1st Qu.:3.248  
#>  Median :2.378   Median :16.06   Median :40.11   Median :3.395  
#>  Mean   :2.399   Mean   :16.05   Mean   :40.10   Mean   :3.407  
#>  3rd Qu.:2.557   3rd Qu.:16.90   3rd Qu.:41.01   3rd Qu.:3.563  
#>  Max.   :2.968   Max.   :20.54   Max.   :43.92   Max.   :3.945  
#>                  NA's   :258                                    
#>       bmi3          hc_CB03m          age4            bmi4      
#>  Min.   :13.52   Min.   :36.94   Min.   :4.008   Min.   :13.65  
#>  1st Qu.:15.62   1st Qu.:40.40   1st Qu.:4.251   1st Qu.:16.22  
#>  Median :16.73   Median :41.42   Median :4.414   Median :17.03  
#>  Mean   :16.67   Mean   :41.43   Mean   :4.429   Mean   :17.01  
#>  3rd Qu.:17.49   3rd Qu.:42.43   3rd Qu.:4.595   3rd Qu.:17.93  
#>  Max.   :20.25   Max.   :45.47   Max.   :4.988   Max.   :21.53  
#>  NA's   :191                                     NA's   :253    
#>     hc_CB04m          age7            bmi7        hc_CB05t10m   
#>  Min.   :37.92   Min.   :5.037   Min.   :13.29   Min.   :39.43  
#>  1st Qu.:41.49   1st Qu.:5.973   1st Qu.:16.30   1st Qu.:42.76  
#>  Median :42.28   Median :6.343   Median :17.37   Median :43.66  
#>  Mean   :42.39   Mean   :6.360   Mean   :17.31   Mean   :43.72  
#>  3rd Qu.:43.27   3rd Qu.:6.757   3rd Qu.:18.28   3rd Qu.:44.64  
#>  Max.   :46.29   Max.   :8.151   Max.   :22.38   Max.   :48.54  
#>                                  NA's   :81                     
#>      age11           bmi11        hc_CB10t13m        age15      
#>  Min.   :10.01   Min.   :13.27   Min.   :41.30   Min.   :13.08  
#>  1st Qu.:10.71   1st Qu.:16.33   1st Qu.:44.93   1st Qu.:14.03  
#>  Median :11.12   Median :17.34   Median :45.84   Median :14.45  
#>  Mean   :11.13   Mean   :17.35   Mean   :45.90   Mean   :14.47  
#>  3rd Qu.:11.51   3rd Qu.:18.32   3rd Qu.:46.90   3rd Qu.:14.90  
#>  Max.   :12.86   Max.   :22.07   Max.   :50.09   Max.   :16.43  
#>                  NA's   :94                                     
#>      bmi15        hc_CB13t17m        age20           bmi20      
#>  Min.   :12.96   Min.   :41.83   Min.   :17.01   Min.   :12.87  
#>  1st Qu.:16.21   1st Qu.:45.78   1st Qu.:18.06   1st Qu.:15.91  
#>  Median :17.20   Median :46.80   Median :18.65   Median :16.90  
#>  Mean   :17.22   Mean   :46.84   Mean   :18.71   Mean   :16.87  
#>  3rd Qu.:18.12   3rd Qu.:47.86   3rd Qu.:19.33   3rd Qu.:17.76  
#>  Max.   :21.88   Max.   :51.46   Max.   :22.02   Max.   :21.25  
#>  NA's   :86                                      NA's   :145    
#>   hc_CB17t23m        age26           bmi26        hc_CB23t29m   
#>  Min.   :42.45   Min.   :23.03   Min.   :12.72   Min.   :44.11  
#>  1st Qu.:46.69   1st Qu.:24.38   1st Qu.:15.62   1st Qu.:47.76  
#>  Median :47.71   Median :25.17   Median :16.46   Median :48.92  
#>  Mean   :47.78   Mean   :25.25   Mean   :16.52   Mean   :48.91  
#>  3rd Qu.:48.88   3rd Qu.:26.06   3rd Qu.:17.38   3rd Qu.:50.09  
#>  Max.   :53.25   Max.   :28.61   Max.   :21.12   Max.   :53.60  
#>                                  NA's   :148                    
#>      age32           bmi32        hc_CB29t35m        age40      
#>  Min.   :29.00   Min.   :12.18   Min.   :44.73   Min.   :35.00  
#>  1st Qu.:30.22   1st Qu.:15.28   1st Qu.:48.49   1st Qu.:36.21  
#>  Median :30.97   Median :16.26   Median :49.62   Median :37.04  
#>  Mean   :31.03   Mean   :16.29   Mean   :49.63   Mean   :37.18  
#>  3rd Qu.:31.73   3rd Qu.:17.17   3rd Qu.:50.77   3rd Qu.:38.02  
#>  Max.   :34.51   Max.   :20.30   Max.   :54.30   Max.   :41.15  
#>                  NA's   :297                                    
#>      bmi40        hc_CB35t44m        age50           bmi50      
#>  Min.   :10.66   Min.   :45.49   Min.   :44.03   Min.   :10.76  
#>  1st Qu.:15.13   1st Qu.:49.11   1st Qu.:45.44   1st Qu.:14.81  
#>  Median :16.05   Median :50.31   Median :46.22   Median :15.71  
#>  Mean   :16.00   Mean   :50.33   Mean   :46.18   Mean   :15.76  
#>  3rd Qu.:16.88   3rd Qu.:51.45   3rd Qu.:46.84   3rd Qu.:16.76  
#>  Max.   :20.41   Max.   :56.37   Max.   :49.29   Max.   :19.62  
#>  NA's   :159                                     NA's   :205    
#>   hc_CB44t56m   
#>  Min.   :45.16  
#>  1st Qu.:49.89  
#>  Median :51.23  
#>  Mean   :51.14  
#>  3rd Qu.:52.32  
#>  Max.   :56.28  
#>