JointAI: Joint Analysis and Imputation of Missing Values

The JointAI package performs simultaneous imputation and inference for incomplete data using the Bayesian framework. Distributions for incomplete variables are specified automatically and modeled jointly with the analysis model.

Main functions

The package has three main functions, lm_imp, glm_imp and lme_imp, that allow analysis using linear regression, generalized linear regression and linear mixed effects models. As far as possible, the specification of the functions is the same as the specification of their complete data versions (lm, glm and lme).

Results can be summarized and printed with summary.JointAI, and visualized using traceplot or densplot.

Evaluation and export

Two criteria for evaluation of convergence and precision of the posterior estimate are available: GR_crit and MC_error

Imputed data can be exported to SPSS data using get_MIdat.

References

Erler, N. S., Rizopoulos, D., Rosmalen, J. V., Jaddoe, V. W., Franco, O. H., & Lesaffre, E. M. (2016). Dealing with missing covariates in epidemiologic studies: A comparison between multiple imputation and a full Bayesian approach. Statistics in Medicine, 35(17), 2955-2974.