Fitted object of class JointAI

An object returned by one of the functions lm_imp(), glm_imp(), clm_imp(), lme_imp(), glme_imp(), clmm_imp(), survreg_imp() or coxph_imp().

Value

analysis_type

lm, glm, clm, lme, glme, clmm, survreg or coxph with attributes family and link

data

the original (incomplete) dataset

models

named vector specifying the models used for longitudinal and incomplete covariates

fixed

supplied fixed effects formula

random

supplied random effects formula

Mlist

a list: containing the data, split up into

outcome (y)

censoring indicator for survival outcomes (cens)

cross-sectional main effects (Xc)

cross-sectional interactions (Xic)

longitudinal main effects (Xl)

longitudinal interactions (Xil)

categorical cross-sectional incomplete variables (Xcat)

categorical longitudinal variables (Xlcat)

transformed cross-sectional variables (Xtrafo)

random effects design matrix (Z)

and other important specifications:

a list naming which columns of the above matrices are covariates in the analysis model (cols_main)

a list giving the names of the covariates in the analysis model per matrix (names_main)

specification for transformations (trafos)

specification for hierarchical centering (hc_list)

reference values and dummies for categorical variables (refs)

vector of auxiliary variables (auxvars)

grouping specification (groups)

the vector of variables to be scaled (scale_vars)

updated fixed effects structure (fixed2)

the number of categories if the outcome of the analysis model is categorical (ncat)

the number of subjects (N)

whether posterior predictive checks are be enabled ppc (not yet used)

whether ridge shrinkage priors should are used for the regression coefficients of the analysis model (ridge)

the number of random effects (nranef)

K

matrix specifying the indices of the regression coefficients that are related to different parts of the model

K_imp

matrix specifying the indices of regression coefficients for the imputation models relating to different covariates

mcmc_settings

a list with elements

modelfile

name and path of JAGS model file

n.chains

number of MCMC chains

n.adapt

number of iterations in the adaptive phase

n.iter

number of iterations in the MCMC sample

variable.names

monitored nodes

thin

thinning of the MCMC sample

inits

a list containing the initial values that were passed to rjags

parallel

whether parallel sampling was used

ncores

how many cores were used in parallel sampling

monitor_params

the list of parameter groups to be monitored

data_list

list with data that was passed to rjags

scale_pars

matrix with parameters used to center and scale the continuous variables

hyperpars

a list containing the values of the hyperparameters used

imp_par_list

a list with parameters used to write the imputation model syntax

model

JAGS model

sample

MCMC sample on the sampling scale (included only if keep_scaled_sample = TRUE)

MCMC

MCMC sample, scaled back to the scale of the data

time

the computational time used for the sampling (adaptive phase + sampling)

call

the original call