Fitted object of class JointAI

An object returned by one of the functions lm_imp(), glm_imp() or lme_imp().

Value

analysis_type

lm, glm or lme, with attributes family and link

data

the original dataset

meth

named vector specifying imputation methods and sequence

fixed

supplied fixed effects structure

random

supplied random effects structure

Mlist

a list of matrices that contain the data split up into outcome (y), cross-sectional main effects (Xc), cross-sectional interactions (Xic), longitudinal main effects (Xl), longitudinal interactions (Xil), categorical incomplete variables (Xcat), transformed cross-sectional variables (Xtrafo), random effects design matrix (Z), the vector of variables to be scaled (scale_vars), reference values and dummies for categorical variables (refs), specification for transformations (trafos), specification for hierarchical centering (hc_list), vector of auxiliary variables (auxvars), grouping specification (groups), updated fixed effects structure (fixed2), names of updated design matrix (X2_names)

refcats

A list naming the reference categories for all categorical covariates.

K

matrix specifying the indices of the regression coefficients that are related to different parts of the model

K_imp

matrix specifying the indices of regression coefficients for the imputation models relating to different covariates

mcmc_settings

a list with elements

MCMCpackage

which package has been used (at the moment only JAGS is implemented)

modelfile

name and path of JAGS model file

n.chains

number of MCMC chains

n.adapt

number of iterations in the adaptive phase

n.iter

number of iterations in the MCMC sample

variable.names

monitored nodes

thin

thinning of the MCMC sample

inits

a list containing the initial values that were used

data_list

list with data that was passed to JAGS

scale_pars

matrix with parameters used to center and scale the continuous variables

hyperpars

a list containing the values of the hyperparameters used

model

JAGS model

sample

MCMC sample (Note: if continuous variables have been scaled during the sampling, the posterior sample here is on the scaled scale, not on the original scale.)

MCMC

if scaling was done: MCMC sample, scaled back to original scale

time

the computational time used for the sampling (adaptive phase + sampling)

call

the original call