Get default values for hyperparameters Prints the list of default values for the hyperparameters

Get default values for hyperparameters Prints the list of default values for the hyperparameters

default_hyperpars(family = "gaussian", link = "identity", nranef = NULL)

Arguments

family
distribution family of the analysis model (gaussian, binomial, poisson or Gamma)

link
link function (if the link is already given in the family, e.g. family = binomial("logit")) this argument does not need to be specified

nranef
number of random effects

Value

A list containing the default hyperparameters for JointAI models. The elements of the list are

analysis_model: hyperparameters for the analysis model

mu_reg_main	mean in the priors for regression coefficients

tau_reg_main	precision in the priors for regression coefficients

a_tau_main	scale parameter in gamma prior for precision of outcome

b_tau_main	rate parameter in gamma prior for precision of outcome

mu_reg_main

Z: hyperparameters for the random effects in mixed models

RinvD	scale matrix in Wishart prior (*) for random effects covariance matrix

KinvD	degrees of freedom in Wishart prior for random effects covariance matrix

a_diag_RinvD	scale parameter in gamma prior for the diagonal elements of RinvD

b_diag_RinvD	rate parameter in gamma prior for the diagonal elements of RinvD

RinvD

(*) when there is only one random effect a gamma distribution is used instead of the Wishart

norm: hyperparameters for normal and lognormal imputation models

mu_reg_norm	mean in the priors for regression coefficients

tau_reg_norm	precision in the priors for regression coefficients

a_tau_norm	scale parameter in gamma prior for precision of imputed variable

b_tau_norm	rate parameter in gamma prior for precision of imputed variable

mu_reg_norm

gamma: hyperparameters for gamma imputation models

mu_reg_gamma	mean in the priors for regression coefficients

tau_reg_gamma	precision in the priors for regression coefficients

a_tau_gamma	scale parameter in gamma prior for precision of imputed variable

b_tau_gamma	rate parameter in gamma prior for precision of imputed variable

mu_reg_gamma

beta: hyperparameters for beta imputation models

mu_reg_beta	mean in the priors for regression coefficients

tau_reg_beta	precision in the priors for regression coefficients

a_tau_beta	scale parameter in gamma prior for precision of imputed variable

b_tau_beta	rate parameter in gamma prior for precision of imputed variable

mu_reg_beta

logit: hyperparameters for logistic imputation models

mu_reg_logit	mean in the priors for regression coefficients

tau_reg_logit	precision in the priors for regression coefficients

mu_reg_logit

multinomial: hyperparameters for multinomial imputation models

mu_reg_multinomial	mean in the priors for regression coefficients

tau_reg_multinomial	precision in the priors for regression coefficients

mu_reg_multinomial

ordinal: hyperparameters for ordinal imputation models

mu_reg_ordinal	mean in the priors for regression coefficients

tau_reg_ordinal	precision in the priors for regression coefficients

mu_delta_ordinal	mean in the prior for the intercepts

tau_delta_ordinal	precision in the priors for the intercepts

mu_reg_ordinal