Get default values for hyperparameters

Prints the list of default values for the hyperparameters.

default_hyperpars()

Details

norm: hyperparameters for normal and lognormal models

mu_reg_norm

mean in the priors for regression coefficients

tau_reg_norm

precision in the priors for regression coefficients

shape_tau_norm

shape parameter in Gamma prior for precision of imputed variable

rate_tau_norm

rate parameter in Gamma prior for precision of imputed variable

gamma: hyperparameters for Gamma models

mu_reg_gamma

mean in the priors for regression coefficients

tau_reg_gamma

precision in the priors for regression coefficients

shape_tau_gamma

shape parameter in Gamma prior for precision of imputed variable

rate_tau_gamma

rate parameter in Gamma prior for precision of imputed variable

beta: hyperparameters for beta models

mu_reg_beta

mean in the priors for regression coefficients

tau_reg_beta

precision in the priors for regression coefficients

shape_tau_beta

shape parameter in Gamma prior for precision of imputed variable

rate_tau_beta

rate parameter in Gamma prior for precision of imputed variable

logit: hyperparameters for logistic models

mu_reg_logit

mean in the priors for regression coefficients

tau_reg_logit

precision in the priors for regression coefficients

probit: hyperparameters for probit models

mu_reg_logit

mean in the priors for regression coefficients

tau_reg_logit

precision in the priors for regression coefficients

multinomial: hyperparameters for multinomial models

mu_reg_multinomial

mean in the priors for regression coefficients

tau_reg_multinomial

precision in the priors for regression coefficients

ordinal: hyperparameters for ordinal models

mu_reg_ordinal

mean in the priors for regression coefficients

tau_reg_ordinal

precision in the priors for regression coefficients

mu_delta_ordinal

mean in the prior for the intercepts

tau_delta_ordinal

precision in the priors for the intercepts

Z: function creating hyperparameters for the random effects in mixed models, with output elements

RinvD

scale matrix in Wishart prior (*) for random effects covariance matrix

KinvD

degrees of freedom in Wishart prior for random effects covariance matrix

shape_diag_RinvD

shape parameter in Gamma prior for the diagonal elements of

RinvD

rate_diag_RinvD

rate parameter in Gamma prior for the diagonal elements of

RinvD

(*) when there is only one random effect a Gamma distribution is used instead of the Wishart and RinvD and KinvD are NULL

surv: parameters for survival models (parametric and proportional hazard)

mu_reg_surv

mean in the priors for regression coefficients

tau_reg_surv

precision in the priors for regression coefficients

coxph: parameters for Cox proportional hazards models

c

confidence in prior guess for the hazard function

r

failure rate per unit time

eps

time increment

Examples

default_hyperpars()

#> $norm
#>    mu_reg_norm   tau_reg_norm shape_tau_norm  rate_tau_norm 
#>          0e+00          1e-04          1e-02          1e-02 
#> 
#> $gamma
#>    mu_reg_gamma   tau_reg_gamma shape_tau_gamma  rate_tau_gamma 
#>           0e+00           1e-04           1e-02           1e-02 
#> 
#> $beta
#>    mu_reg_beta   tau_reg_beta shape_tau_beta  rate_tau_beta 
#>          0e+00          1e-04          1e-02          1e-02 
#> 
#> $logit
#>  mu_reg_logit tau_reg_logit 
#>         0e+00         1e-04 
#> 
#> $poisson
#>  mu_reg_poisson tau_reg_poisson 
#>           0e+00           1e-04 
#> 
#> $probit
#>  mu_reg_probit tau_reg_probit 
#>          0e+00          1e-04 
#> 
#> $multinomial
#>  mu_reg_multinomial tau_reg_multinomial 
#>               0e+00               1e-04 
#> 
#> $ordinal
#>    mu_reg_ordinal   tau_reg_ordinal  mu_delta_ordinal tau_delta_ordinal 
#>             0e+00             1e-04             0e+00             1e-04 
#> 
#> $Z
#> function(nranef) {
#>       if (nranef > 1) {
#>         RinvD <- diag(as.numeric(rep(NA, nranef)))
#>         KinvD <- nranef
#>       } else {
#>         RinvD <- KinvD <- NULL
#>       }
#> 
#>       list(
#>         RinvD = RinvD,
#>         KinvD = KinvD,
#>         shape_diag_RinvD = 0.1,
#>         rate_diag_RinvD = 0.01
#>       )
#>     }
#> <bytecode: 0x00000000065eb108>
#> <environment: 0x0000000006923cc0>
#> 
#> $surv
#>  mu_reg_surv tau_reg_surv 
#>        0.000        0.001 
#> 
#> $coxph
#>     c     r   eps 
#> 1e-03 1e-01 1e-10 
#>