Joint analysis and imputation of incomplete data

lm_imp, glm_imp and lme_imp estimate linear, generalized linear and linear mixed models, respectively, using MCMC sampling.

lm_imp(formula, data, n.chains = 3, n.adapt = 100, n.iter = 0, thin = 1,
  monitor_params = NULL, inits = TRUE, modelname = NULL,
  modeldir = NULL, overwrite = NULL, keep_model = FALSE, quiet = TRUE,
  progress.bar = "text", warn = TRUE, mess = TRUE, auxvars = NULL,
  meth = NULL, refcats = NULL, scale_vars = NULL, hyperpars = NULL, ...)

glm_imp(formula, family, data, n.chains = 3, n.adapt = 100, n.iter = 0,
  thin = 1, monitor_params = NULL, inits = TRUE, modelname = NULL,
  modeldir = NULL, overwrite = NULL, keep_model = FALSE, quiet = TRUE,
  progress.bar = "text", warn = TRUE, mess = TRUE, auxvars = NULL,
  meth = NULL, refcats = NULL, scale_vars = NULL, hyperpars = NULL, ...)

lme_imp(fixed, data, random, n.chains = 3, n.adapt = 100, n.iter = 0,
  thin = 1, monitor_params = NULL, inits = TRUE, modelname = NULL,
  modeldir = NULL, overwrite = NULL, keep_model = FALSE, quiet = TRUE,
  progress.bar = "text", warn = TRUE, mess = TRUE, auxvars = NULL,
  meth = NULL, refcats = NULL, scale_vars = NULL, hyperpars = NULL, ...)

Arguments

formula
a two sided model formula (see formula)

data
a data frame

n.chains
the number of parallel chains for the model

n.adapt
the number of iterations for adaptation. See adapt for details. If n.adapt = 0 then no adaptation takes place.

n.iter
number of iterations to monitor

thin
thinning interval for monitors

monitor_params
named vector specifying which parameters should be monitored, see details.

inits
optional specification of initial values in the form of a list or a function (see jags.model. If omitted, initial values will be generated automatically. It is an error to supply an initial value for an observed node.

modelname
optional; character string specifying the name of model file (including the ending, either .R or .txt). If unspecified a random name will be generated.

modeldir
optional; directory containing model file. If unspecified a temporary directory will be created.

overwrite
logical; whether an existing model file with the specified modeldir/modelname should be overwritten. If set to FALSE (default) and a model already exists, that model will be used.

keep_model
logical; whether the created JAGS model should be saved or removed from the disk (FALSE; default) when the sampling has finished.

quiet
if TRUE then messages generated during compilation will be suppressed, as well as the progress bar during adaptation.

progress.bar
type of progress bar. Possible values are "text", "gui", and "none". See Details.

warn
logical; should warnings be given?; default is TRUE. Note: this applies only to warnings given directly by JointAI.

mess
logical; should messages be given?; default is TRUE. Note: this applies only to messages given directly by JointAI.

auxvars
optional vector of variable names that should be used as predictors in the imputation procedure (and will be imputed if necessary) but are not part of the analysis model

meth
optional named vector specifying imputation model types and order. If NULL (default) imputation models will be determined automatically based on the class of the columns of data that contain missing values (see Details). The default order is according to the proportion of missing values (increasing).

refcats
optional; a named list specifying which category should be used as reference category for each of the categorical variables. Options are the category label, the category number, 'first" (the first category) or "largest" (chooses the category with the most observations). Default is "first".

scale_vars
optional; named vector of (continuous) variables that will be scaled (such that mean = 0 and sd = 1) to improve convergence of the MCMC sampling. Default is that all continuous variables that are not transformed by a function (e.g. log(), ns()) will be scaled. Variables for which "lognorm" is set as imputation model are only scaled with regards to the standard deviation, but not centered. If set to FALSE no scaling will be done.

hyperpars
list of hyperparameters, as obtained by default_hyperpars(); only needs to be supplied if hyperparameters other than the default should be used

...
additional, optional arguments, see below

family
only for glm_imp: a description of the distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function. (See family and the `Details` section below.)

fixed
a two sided formula describing the fixed-effects part of the model (see formula)

random
only for lme_imp: a one-sided formula of the form ~x1 + ... + xn | g, where x1 + ... + xn specifies the model for the random effects and g the grouping variable

Value

An object of class JointAI

Optional arguments

There are some optional parameters that can be passed to …

scale_pars

optional matrix of parameters used for centering and scaling continuous covariates. If not specified, this will be calculated automatically. If FALSE, no scaling will be done.

Details

Implemented distribution families and link functions for glm_imp()

gaussian	with links: identity, log

binomial	with links: logit, probit, log, cloglog

Gamma	with links: identity, log

poisson	with links: log, identity

gaussian

Imputation methods
Implemented imputation models that can be chosen in the argument meth are:

norm	linear model

lognorm	log-linear model for skewed continuous data

gamma	gamma model (with log-link) for skewed continuous data

beta	beta model (with logit-link) for skewed continuous data in (0, 1)

logit	logistic model for binary data

multilogit	multinomial logit model for unordered categorical variables

cumlogit	cumulative logit model for ordered categorical variables

norm

Parameters to follow (monitor_params)
Named vector specifying which parameters should be monitored. This can be done either directly by specifying the name of the parameter or indirectly by one of the key words summarizing a number of parameters. Except for other, in which parameter names are specified directly, parameter (groups) are just set as TRUE or FALSE. If left unspecified, monitor_params = c("analysis_main" = TRUE) will be used.

name/key word	what is monitored

analysis_main	betas, tau_y and sigma_y

analysis_random	ranef, D, invD, RinvD

imp_pars	alphas, tau_imp, gamma_imp, delta_imp

imps	imputed values

betas	regression coefficients of the analysis model

tau_y	precision of the residuals from the analysis model

sigma_y	standard deviation of the residuals from the analysis model

ranef	random effects

D	covariance matrix of the random effects

invD	inverse of D

RinvD	matrix in prior for invD

alphas	regression coefficients in the imputation models

tau_imp	precision parameters of the residuals from imputation models

gamma_imp	intercepts in ordinal imputation models

delta_imp	increments of ordinal intercepts

other	additional parameters

name/key word

For example: monitor_params = c("analysis_main" = TRUE, "tau_y" = FALSE) would monitor the regression parameters betas and residual standard deviation sigma_y, but not the residual precision. monitor_params = c(imps = TRUE) would monitor betas, tau_y, and sigma_y (because analysis_main = TRUE by default) as well as the imputed values.
See also

Examples

mod1 <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)

#> This is new software. Please report any bug to the package maintainer.

mod2 <- glm_imp(B1 ~ C1 + C2 + M2, data = wideDF,
                family = binomial(link = "logit"), n.iter = 100)

#> This is new software. Please report any bug to the package maintainer.

mod3 <- lme_imp(y ~ C1 + B2 + L1 + time, random = ~ time|id,
                data = longDF, n.iter = 500)

#> This is new software. Please report any bug to the package maintainer.