Create a new dataframe for prediction

Build a data.frame for prediction, where one variable varies and all other variables are set to the reference value (median for continuous variables).

predDF(...)

# S3 method for formula
predDF(formula, dat, var, ...)

# S3 method for JointAI
predDF(object, var, ...)

Arguments

...
optional, additional arguments (currently not used)

formula
a two sided model formula (see formula)

dat
original data

var
name of variable that should be varying

object
object inheriting from class "JointAI"

See also

Examples

# fit a JointAI model
mod <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)

#> This is new software. Please report any bug to the package maintainer.

# generate a dataframe with varying "C2" and reference values for all other variables in the model
newDF <- predDF(mod, var = "C2")

head(newDF)

#>           y       C1         C2 M2
#> 1 -3.183137 1.434298 -0.9226220  1
#> 2 -3.183137 1.434298 -0.9046136  1
#> 3 -3.183137 1.434298 -0.8866052  1
#> 4 -3.183137 1.434298 -0.8685968  1
#> 5 -3.183137 1.434298 -0.8505885  1
#> 6 -3.183137 1.434298 -0.8325801  1