Simulated Longitudinal Data in Long and Wide Format

This data was simulated to mimic data from a longitudinal cohort study following mothers and their child from birth until approximately 4 years of age. It contains 3903 observations of 499 mother-child pairs. Children's BMI and head circumference was measured repeatedly and their age in months was recorded at each measurement. Furthermore, the data contain several baseline variables with information on the mothers' demographics and socioeconomic status.

simLong

simWide

Format

simLong: A data frame in long format with 3894 rows and 13 variables

simWide: A data frame in wide format with 499 rows and 45 variables

Baseline covariates

(in simLong and simWide)

GESTBIR

gestational age at birth (in weeks)

ETHN

ethnicity (binary: European vs. other)

AGE_M

age of the mother at intake

HEIGHT_M

height of the mother (in cm)

PARITY

number of times the mother has given birth (binary: 0 vs. >=1)

SMOKE

smoking status of the mother during pregnancy (3 ordered categories: never smoked during pregnancy, smoked until pregnancy was known, continued smoking in pregnancy)

EDUC

educational level of the mother (3 ordered categories: low, mid, high)

MARITAL

marital status (3 categories)

ID

subject identifier

Long-format variables

(only in simLong)

time

measurement occasion/visit (by design, children should be measured at/around 1, 2, 3, 4, 7, 11, 15, 20, 26, 32, 40 and 50 months of age)

age

child age at measurement time in months

bmi

child BMI

hc

child head circumference in cm

Wide-format variables

(only in simWide)

age1, age2, age3, age4, age7, age11, age15, age20, age26, age32, age40, age50

child age at the repeated measurements in months

bmi1, bmi2, bmi3, bmi4, bmi7, bmi11, bmi15, bmi20, bmi26, bmi32, bmi40, bmi50

repeated measurements of child BMI

age1, age2, age3, age4, age7, age11, age15, age20, age26, age32, age40, age50

repeated measurements of child head circumference in cm

Examples

 summary(simLong)

#>     GESTBIR            ETHN          AGE_M          HEIGHT_M      PARITY    
#>  Min.   :36.29   European:2448   Min.   :17.06   Min.   :147.0   0   :2332  
#>  1st Qu.:39.37   other   :1355   1st Qu.:26.92   1st Qu.:162.9   >= 1:1469  
#>  Median :40.34   NA's    :  91   Median :30.34   Median :168.0   NA's:  93  
#>  Mean   :40.24                   Mean   :30.52   Mean   :167.9              
#>  3rd Qu.:41.13                   3rd Qu.:34.17   3rd Qu.:172.9              
#>  Max.   :43.79                   Max.   :46.73   Max.   :192.5              
#>                                                  NA's   :76                 
#>                               SMOKE        EDUC                 MARITAL    
#>  never smoked during pregnancy   :2595   high:1740   married        :1782  
#>  smoked until pregnancy was known: 276   mid :1544   living together:1421  
#>  continued smoking in pregnancy  : 565   low : 346   no partner     : 453  
#>  NA's                            : 458   NA's: 264   NA's           : 238  
#>                                                                            
#>                                                                            
#>                                                                            
#>       time            age               bmi              hc       
#>  Min.   : 1.00   Min.   : 0.3119   Min.   :10.66   Min.   :32.58  
#>  1st Qu.: 4.00   1st Qu.: 4.3629   1st Qu.:15.52   1st Qu.:42.38  
#>  Median :13.00   Median :12.8593   Median :16.61   Median :46.03  
#>  Mean   :17.23   Mean   :16.4314   Mean   :16.58   Mean   :45.55  
#>  3rd Qu.:26.00   3rd Qu.:25.4791   3rd Qu.:17.65   3rd Qu.:48.93  
#>  Max.   :50.00   Max.   :49.2852   Max.   :22.38   Max.   :56.37  
#>                                                                   
#>       ID           
#>  Length:3894       
#>  Class :character  
#>  Mode  :character  
#>                    
#>                    
#>                    
#>                    

 summary(simWide)

#>     GESTBIR            ETHN         AGE_M          HEIGHT_M      PARITY   
#>  Min.   :36.29   European:304   Min.   :17.06   Min.   :147.0   0   :305  
#>  1st Qu.:39.32   other   :181   1st Qu.:26.53   1st Qu.:162.6   >= 1:182  
#>  Median :40.30   NA's    : 14   Median :30.04   Median :167.9   NA's: 12  
#>  Mean   :40.19                  Mean   :30.24   Mean   :167.7             
#>  3rd Qu.:41.10                  3rd Qu.:33.86   3rd Qu.:172.6             
#>  Max.   :43.79                  Max.   :46.73   Max.   :192.5             
#>                                                 NA's   :10                
#>                               SMOKE       EDUC                MARITAL   
#>  never smoked during pregnancy   :332   high:218   married        :224  
#>  smoked until pregnancy was known: 34   mid :196   living together:180  
#>  continued smoking in pregnancy  : 73   low : 46   no partner     : 58  
#>  NA's                            : 60   NA's: 39   NA's           : 37  
#>                                                                         
#>                                                                         
#>                                                                         
#>       ID                 age1             bmi1         hc_CB0t2m    
#>  Length:499         Min.   :0.3119   Min.   :11.32   Min.   :32.35  
#>  Class :character   1st Qu.:0.9581   1st Qu.:14.11   1st Qu.:36.12  
#>  Mode  :character   Median :1.1191   Median :15.16   Median :37.16  
#>                     Mean   :1.1238   Mean   :15.12   Mean   :37.21  
#>                     3rd Qu.:1.3020   3rd Qu.:16.09   3rd Qu.:38.23  
#>                     Max.   :1.7615   Max.   :19.43   Max.   :41.53  
#>                                      NA's   :175                    
#>       age2            bmi2          hc_CB02m          age3      
#>  Min.   :2.002   Min.   :11.72   Min.   :35.93   Min.   :3.001  
#>  1st Qu.:2.215   1st Qu.:15.31   1st Qu.:39.25   1st Qu.:3.248  
#>  Median :2.378   Median :16.09   Median :40.11   Median :3.397  
#>  Mean   :2.399   Mean   :16.07   Mean   :40.10   Mean   :3.408  
#>  3rd Qu.:2.556   3rd Qu.:16.90   3rd Qu.:41.02   3rd Qu.:3.563  
#>  Max.   :2.968   Max.   :20.54   Max.   :43.92   Max.   :3.945  
#>                  NA's   :258                                    
#>       bmi3          hc_CB03m          age4            bmi4      
#>  Min.   :13.52   Min.   :36.94   Min.   :4.008   Min.   :13.59  
#>  1st Qu.:15.62   1st Qu.:40.41   1st Qu.:4.251   1st Qu.:16.23  
#>  Median :16.73   Median :41.43   Median :4.414   Median :17.04  
#>  Mean   :16.67   Mean   :41.43   Mean   :4.429   Mean   :17.02  
#>  3rd Qu.:17.50   3rd Qu.:42.43   3rd Qu.:4.595   3rd Qu.:17.95  
#>  Max.   :20.25   Max.   :45.48   Max.   :4.988   Max.   :21.53  
#>  NA's   :191                                     NA's   :249    
#>     hc_CB04m          age7            bmi7        hc_CB05t10m   
#>  Min.   :37.92   Min.   :5.037   Min.   :13.29   Min.   :39.43  
#>  1st Qu.:41.49   1st Qu.:5.973   1st Qu.:16.30   1st Qu.:42.75  
#>  Median :42.29   Median :6.341   Median :17.37   Median :43.65  
#>  Mean   :42.39   Mean   :6.358   Mean   :17.31   Mean   :43.72  
#>  3rd Qu.:43.27   3rd Qu.:6.755   3rd Qu.:18.28   3rd Qu.:44.64  
#>  Max.   :46.29   Max.   :8.151   Max.   :22.38   Max.   :48.54  
#>                                  NA's   :80                     
#>      age11           bmi11        hc_CB10t13m        age15      
#>  Min.   :10.01   Min.   :13.27   Min.   :41.30   Min.   :13.08  
#>  1st Qu.:10.72   1st Qu.:16.35   1st Qu.:44.94   1st Qu.:14.04  
#>  Median :11.13   Median :17.34   Median :45.85   Median :14.45  
#>  Mean   :11.14   Mean   :17.36   Mean   :45.90   Mean   :14.47  
#>  3rd Qu.:11.51   3rd Qu.:18.33   3rd Qu.:46.90   3rd Qu.:14.90  
#>  Max.   :12.86   Max.   :22.07   Max.   :50.09   Max.   :16.43  
#>                  NA's   :94                                     
#>      bmi15        hc_CB13t17m        age20           bmi20      
#>  Min.   :12.96   Min.   :41.83   Min.   :17.01   Min.   :12.87  
#>  1st Qu.:16.20   1st Qu.:45.78   1st Qu.:18.07   1st Qu.:15.92  
#>  Median :17.20   Median :46.78   Median :18.65   Median :16.90  
#>  Mean   :17.22   Mean   :46.84   Mean   :18.71   Mean   :16.87  
#>  3rd Qu.:18.13   3rd Qu.:47.85   3rd Qu.:19.33   3rd Qu.:17.77  
#>  Max.   :21.88   Max.   :51.46   Max.   :22.02   Max.   :21.25  
#>  NA's   :84                                      NA's   :149    
#>   hc_CB17t23m        age26           bmi26        hc_CB23t29m   
#>  Min.   :42.45   Min.   :23.03   Min.   :12.72   Min.   :44.11  
#>  1st Qu.:46.68   1st Qu.:24.38   1st Qu.:15.62   1st Qu.:47.77  
#>  Median :47.74   Median :25.16   Median :16.47   Median :48.93  
#>  Mean   :47.78   Mean   :25.25   Mean   :16.51   Mean   :48.91  
#>  3rd Qu.:48.87   3rd Qu.:26.06   3rd Qu.:17.38   3rd Qu.:50.09  
#>  Max.   :53.25   Max.   :28.61   Max.   :21.12   Max.   :53.60  
#>                                  NA's   :150                    
#>      age32           bmi32        hc_CB29t35m        age40      
#>  Min.   :29.00   Min.   :12.18   Min.   :44.73   Min.   :35.00  
#>  1st Qu.:30.22   1st Qu.:15.26   1st Qu.:48.49   1st Qu.:36.20  
#>  Median :30.97   Median :16.22   Median :49.61   Median :37.03  
#>  Mean   :31.03   Mean   :16.28   Mean   :49.63   Mean   :37.18  
#>  3rd Qu.:31.74   3rd Qu.:17.19   3rd Qu.:50.77   3rd Qu.:38.01  
#>  Max.   :34.51   Max.   :20.30   Max.   :54.30   Max.   :41.15  
#>                  NA's   :300                                    
#>      bmi40        hc_CB35t44m        age50           bmi50      
#>  Min.   :10.66   Min.   :45.49   Min.   :44.03   Min.   :10.76  
#>  1st Qu.:15.13   1st Qu.:49.11   1st Qu.:45.43   1st Qu.:14.81  
#>  Median :16.03   Median :50.31   Median :46.22   Median :15.71  
#>  Mean   :15.99   Mean   :50.33   Mean   :46.18   Mean   :15.76  
#>  3rd Qu.:16.87   3rd Qu.:51.46   3rd Qu.:46.84   3rd Qu.:16.77  
#>  Max.   :20.11   Max.   :56.37   Max.   :49.29   Max.   :19.62  
#>  NA's   :158                                     NA's   :206    
#>   hc_CB44t56m   
#>  Min.   :45.16  
#>  1st Qu.:49.90  
#>  Median :51.23  
#>  Mean   :51.14  
#>  3rd Qu.:52.33  
#>  Max.   :56.28  
#>