Summary of an object of class JointAI

Obtain and print the summary, (fixed effects) coefficients (coef) and credible interval (confint) for an object of class 'JointAI'.

# S3 method for JointAI
summary(object, start = NULL, end = NULL,
  thin = NULL, quantiles = c(0.025, 0.975), subset = NULL,
  warn = TRUE, mess = TRUE, ...)

# S3 method for summary.JointAI
print(x, digits = max(3, .Options$digits - 4),
  ...)

# S3 method for JointAI
coef(object, start = NULL, end = NULL, thin = NULL,
  subset = NULL, warn = TRUE, mess = TRUE, ...)

# S3 method for JointAI
confint(object, parm = NULL, level = 0.95,
  quantiles = NULL, start = NULL, end = NULL, thin = NULL,
  subset = NULL, warn = TRUE, mess = TRUE, ...)

# S3 method for JointAI
print(x, digits = max(4, getOption("digits") - 4), ...)

Arguments

object
object inheriting from class 'JointAI'

start
the first iteration of interest (see window.mcmc)

end
the last iteration of interest (see window.mcmc)

thin
thinning interval (see window.mcmc)

quantiles
posterior quantiles

subset
subset of parameters/variables/nodes (columns in the MCMC sample). Uses the same logic as the argument monitor_params in lm_imp, glm_imp, clm_imp, lme_imp, glme_imp, survreg_imp and coxph_imp.

warn
logical; should warnings be given? Default is TRUE. Note: this applies only to warnings given directly by JointAI.

mess
logical; should messages be given? Default is TRUE. Note: this applies only to messages given directly by JointAI.

…
currently not used

x
an object of class summary.JointAI or JointAI

digits
minimal number of significant digits, see print.default.

parm
same as subset

level
confidence level (default is 0.95)

See also

Examples

mod1 <- lm_imp(y~C1 + C2 + M2, data = wideDF, n.iter = 100)

#> This is new software. Please report any bugs to the package maintainer.

summary(mod1)

#> 
#>  Linear model fitted with JointAI 
#> 
#> Call:
#> lm_imp(formula = y ~ C1 + C2 + M2, data = wideDF, n.iter = 100)
#> 
#> Posterior summary:
#>                 Mean     SD    2.5%   97.5% tail-prob. GR-crit
#> (Intercept)  35.9266 25.078 -13.167 85.0785     0.1667    1.02
#> C1          -27.0295 17.506 -61.589  7.2831     0.1467    1.02
#> C2            0.8441  0.635  -0.335  2.0417     0.1933    1.03
#> M22          -0.4536  0.680  -1.779  0.7489     0.5467    1.00
#> M23           0.0231  0.630  -1.232  1.1619     0.9667    1.06
#> M24          -1.1790  0.680  -2.572  0.0179     0.0667    1.01
#> 
#> Posterior summary of residual std. deviation:
#>         Mean    SD 2.5% 97.5% GR-crit
#> sigma_y 2.24 0.165 1.94  2.59    1.05
#> 
#> 
#> MCMC settings:
#> Iterations = 101:200
#> Sample size per chain = 100 
#> Thinning interval = 1 
#> Number of chains = 3 
#> 
#> Number of observations: 100 

coef(mod1)

#>  (Intercept)           C1           C2          M22          M23          M24 
#>  35.92662356 -27.02953400   0.84411974  -0.45364779   0.02310191  -1.17901466 

confint(mod1)

#>                    2.5%       97.5%
#> (Intercept) -13.1672406 85.07847609
#> C1          -61.5887206  7.28310117
#> C2           -0.3350643  2.04173199
#> M22          -1.7793492  0.74889976
#> M23          -1.2324534  1.16193443
#> M24          -2.5724005  0.01788788
#> sigma_y       1.9425775  2.58765691