Summary of an object of class JointAI

Obtain and print the summary, (fixed effects) coefficients (coef) and credible interval (confint) for an object of class 'JointAI'.

# S3 method for JointAI
summary(object, start = NULL, end = NULL,
  thin = NULL, quantiles = c(0.025, 0.975), subset = NULL,
  exclude_chains = NULL, warn = TRUE, mess = TRUE, ...)

# S3 method for summary.JointAI
print(x, digits = max(3, .Options$digits - 4),
  ...)

# S3 method for JointAI
coef(object, start = NULL, end = NULL, thin = NULL,
  subset = NULL, exclude_chains = NULL, warn = TRUE, mess = TRUE,
  ...)

# S3 method for JointAI
confint(object, parm = NULL, level = 0.95,
  quantiles = NULL, start = NULL, end = NULL, thin = NULL,
  subset = NULL, exclude_chains = NULL, warn = TRUE, mess = TRUE,
  ...)

# S3 method for JointAI
print(x, digits = max(4, getOption("digits") - 4), ...)

Arguments

object
object inheriting from class 'JointAI'

start
the first iteration of interest (see window.mcmc)

end
the last iteration of interest (see window.mcmc)

thin
thinning interval (see window.mcmc)

quantiles
posterior quantiles

subset
subset of parameters/variables/nodes (columns in the MCMC sample). Uses the same logic as the argument monitor_params in *_imp.

exclude_chains
optional vector of the index numbers of chains that should be excluded

warn
logical; should warnings be given? Default is TRUE. (Note: this applies only to warnings given directly by JointAI.)

mess
logical; should messages be given? Default is TRUE. (Note: this applies only to messages given directly by JointAI.)

…
currently not used

x
an object of class summary.JointAI or JointAI

digits
minimal number of significant digits, see print.default.

parm
same as subset

level
confidence level (default is 0.95)

See also

Examples

mod1 <- lm_imp(y ~ C1 + C2 + M2, data = wideDF, n.iter = 100)

summary(mod1)

#> 
#>  Linear model fitted with JointAI 
#> 
#> Call:
#> lm_imp(formula = y ~ C1 + C2 + M2, data = wideDF, n.iter = 100)
#> 
#> Posterior summary:
#>                 Mean     SD   2.5%   97.5% tail-prob. GR-crit
#> (Intercept)  36.7606 25.327  -9.00 85.2688     0.1067   0.999
#> C1          -27.5574 17.701 -61.71  4.4350     0.0867   1.000
#> C2            0.8697  0.672  -0.46  2.1218     0.2200   1.046
#> M22          -0.5440  0.709  -2.11  0.6653     0.4533   1.133
#> M23          -0.0524  0.695  -1.60  1.2580     0.9400   1.082
#> M24          -1.3314  0.739  -3.02 -0.0107     0.0467   1.081
#> 
#> Posterior summary of residual std. deviation:
#>         Mean    SD 2.5% 97.5% GR-crit
#> sigma_y 2.22 0.156 1.94  2.52    1.01
#> 
#> 
#> MCMC settings:
#> Iterations = 101:200
#> Sample size per chain = 100 
#> Thinning interval = 1 
#> Number of chains = 3 
#> 
#> Number of observations: 100 

coef(mod1)

#>  (Intercept)           C1           C2          M22          M23          M24 
#>  36.76060840 -27.55739397   0.86968248  -0.54402053  -0.05237732  -1.33143903 

confint(mod1)

#>                    2.5%       97.5%
#> (Intercept)  -9.0001201 85.26880648
#> C1          -61.7144337  4.43501973
#> C2           -0.4600706  2.12179023
#> M22          -2.1100542  0.66526684
#> M23          -1.5981172  1.25795410
#> M24          -3.0246446 -0.01072354
#> sigma_y       1.9391675  2.51582603