Summarize the results from an object of class JointAI

Obtain and print the summary, (fixed effects) coefficients (coef) and credible interval (confint) for an object of class 'JointAI'.

# S3 method for JointAI
summary(object, start = NULL, end = NULL, thin = NULL,
  quantiles = c(0.025, 0.975), subset = NULL, exclude_chains = NULL,
  outcome = NULL, missinfo = FALSE, warn = TRUE, mess = TRUE, ...)

# S3 method for summary.JointAI
print(x, digits = max(3, .Options$digits - 4), ...)

# S3 method for JointAI
coef(object, start = NULL, end = NULL, thin = NULL,
  subset = NULL, exclude_chains = NULL, warn = TRUE, mess = TRUE, ...)

# S3 method for JointAI
confint(object, parm = NULL, level = 0.95,
  quantiles = NULL, start = NULL, end = NULL, thin = NULL,
  subset = NULL, exclude_chains = NULL, warn = TRUE, mess = TRUE, ...)

# S3 method for JointAI
print(x, digits = max(4, getOption("digits") - 4), ...)

Arguments

object
object inheriting from class 'JointAI'

start
the first iteration of interest (see window.mcmc)

end
the last iteration of interest (see window.mcmc)

thin
thinning interval (integer; see window.mcmc). For example, thin = 1 (default) will keep the MCMC samples from all iterations; thin = 5 would only keep every 5th iteration.

quantiles
posterior quantiles

subset
subset of parameters/variables/nodes (columns in the MCMC sample). Follows the same principle as the argument monitor_params in *_imp.

exclude_chains
optional vector of the index numbers of chains that should be excluded

outcome
optional; vector identifying for which outcomes the summary should be given, either by specifying their indices, or their names (LHS of the respective model formulas as character string).

missinfo
logical; should information on the number and proportion of missing values be included in the summary?

warn
logical; should warnings be given? Default is TRUE.

mess
logical; should messages be given? Default is TRUE.

...
currently not used

x
an object of class summary.JointAI or JointAI

digits
minimal number of significant digits, see print.default.

parm
same as subset (for consistency with confint method for other types of objects)

level
confidence level (default is 0.95)

See also

Examples

mod1 <- lm_imp(y ~ C1 + C2 + M2, data = wideDF, n.iter = 100)

summary(mod1, missinfo = TRUE)

#> 
#> Bayesian linear model fitted with JointAI
#> 
#> Call:
#> lm_imp(formula = y ~ C1 + C2 + M2, data = wideDF, n.iter = 100)
#> 
#> 
#> Posterior summary:
#>                 Mean     SD    2.5%   97.5% tail-prob. GR-crit MCE/SD
#> (Intercept)  37.6037 24.507 -11.124 85.1958     0.1333   1.090 0.0543
#> C1          -28.2447 17.107 -61.349  5.6636     0.1133   1.092 0.0542
#> C2            0.9333  0.651  -0.377  2.1183     0.1867   1.019 0.0654
#> M22          -0.3471  0.668  -1.645  0.9416     0.5867   1.002 0.0577
#> M23           0.0956  0.653  -1.139  1.3270     0.8933   1.047 0.0577
#> M24          -1.1150  0.661  -2.297  0.0627     0.0867   0.997 0.0577
#> 
#> 
#> MCMC settings:
#> Iterations = 101:200
#> Sample size per chain = 100 
#> Thinning interval = 1 
#> Number of chains = 3 
#> 
#> Number of observations: 100 
#> 
#> 
#> Number and proportion of complete cases:
#>         #  %
#> lvlone 93 93
#> 
#> Number and proportion of missing values:
#>    # NA % NA
#> y     0    0
#> C1    0    0
#> M2    3    3
#> C2    4    4
#> 

coef(mod1)

#> $y
#>  (Intercept)           C1           C2          M22          M23          M24 
#>  37.60368434 -28.24471802   0.93325237  -0.34711392   0.09557549  -1.11501623 
#>      sigma_y 
#>   2.21883372 
#> 

confint(mod1)

#> $y
#>                    2.5%      97.5%
#> (Intercept) -11.1236073 85.1958002
#> C1          -61.3494500  5.6635787
#> C2           -0.3768932  2.1182763
#> M22          -1.6449690  0.9415690
#> M23          -1.1386809  1.3270273
#> M24          -2.2967594  0.0626675
#> sigma_y       1.9343879  2.5266842
#>